Prevent flicker from tagmanager setup.
